Move the constraint part of trajectoryConstraints rendering into a separate class
also make the paths update without reloading the whole pathgui
